Data type conv

+++ pet-bokeh/deeplab.py
@@ -34,7 +34,7 @@ class DeepLabModel(object):
         target_size = (self.input_size, self.input_size)
         resized_image = cv2.resize(image, target_size, interpolation=cv2.INTER_AREA)
         #We run the model and get the segmentation results
-        batch_seg_map = self.predict(resized_image[np.newaxis, :, :, :])
+        batch_seg_map = self.predict(resized_image[np.newaxis, :, :, :].astype(np.float32))
         #We apply softmax on a pixel-wise basis
         seg_map = softmax(batch_seg_map[0][:target_size[1], :target_size[0]], axis=-1)
         #We return the channel corresponding to the human segmentation
